# swagger.yml
openapi: 3.0.2
info:
  title: RecipeBook
  version: 1.0.0
  description: 'Simple RecipeBook app where you can create, view, edit and delete your recipes.'
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
paths:
  /users:
    summary: Path used to manage the list of users.
    description: >-
      The REST endpoint/path used to list and create zero or more `User` entities.  This path contains a
      `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `User` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            examples:
              User:
                value:
                  email: test.test@gmail.com
                  password: Passw0rd
                  username: Test
        required: true
      tags:
        - /users
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Email and password are required:
                  value: Email and password are required
                Password must be at least 8 characters long:
                  value: Password must be at least 8 characters long
                Email must be in a valid format:
                  value: Email must be in a valid format
                'Invalid email: {reason}':
                  value: 'Invalid email: {reason}'
                Username must be at least 3 characters long:
                  value: Username must be at least 3 characters long
                Username must be at most 20 characters long:
                  value: Username must be at most 20 characters long
          description: Client side error
        '409':
          content:
            text/html:
              examples:
                Email already exists:
                  value: Email already exists
          description: The email address provided already exists in the system.
      operationId: createUser
      summary: Create a User
      description: Creates a new instance of a `User`.
  /sessions:
    summary: Path used to manage the list of sessions.
    description: >-
      The REST endpoint/path used to list and create zero or more `Session` entities.  This path
      contains a `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `Session` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
            examples:
              Credentials:
                value:
                  email: admin
                  password: admin
        required: true
      tags:
        - /sessions
      responses:
        '201':
          content:
            text/html:
              examples:
                Session:
                  value:
                    id: 330d67de-432b-48ce-a617-0bd6774ebcee
                    userId: 1
                    username: Admin
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Email and password are required:
                  value: Email and password are required
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Invalid password:
                  value: Invalid password
          description: Unauthorized
        '404':
          content:
            text/html:
              examples:
                User not found:
                  value: User not found
          description: Not Found
        '500':
          content:
            text/html:
              examples:
                Internal server error:
                  value: Internal server error
          description: Internal Server Error
      operationId: createSession
      summary: Create a Session
      description: 'Creates a new instance of a `Session`, signing the user in. Returns session ID.'
    delete:
      tags:
        - /sessions
      responses:
        '204':
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
      security:
        -
          Bearer: []
      summary: Destroy a Session
      description: 'Destroys a `Session`, signing the user out. Reads the session ID from Authorization header.'
  /recipes:
    summary: Path used to manage the list of recipes.
    description: >-
      The REST endpoint/path used to list and create zero or more `Recipe` entities.  This path contains
      a `GET` and `POST` operation to perform the list and create recipes, respectively.
    get:
      tags:
        - /recipes
      responses:
        '200':
          content:
            text/html:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
              examples:
                Recipe:
                  value:
                    -
                      id: 3
                      title: Title
                      content: Content
                      userId: 1
          description: Successful response - returns an array of `Recipes` entities.
        '400':
          content:
            text/html:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
      deprecated: false
      security:
        -
          Bearer: []
      operationId: getRecipes
      summary: List All Recipes
      description: Gets a list of all `Recipes` entities.
    post:
      requestBody:
        description: A new `Recipes` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
            examples:
              Recipe:
                value:
                  content: some text
                  title: some text
        required: true
      tags:
        - /recipes
      responses:
        '201':
          content:
            text/html:
              examples:
                Recipe:
                  value:
                    id: 3
                    title: Title
                    content: Content
                    userId: 1
          description: Successful response.
        '400':
          content:
            text/html:
              examples:
                Title and content are required:
                  value: Title and content are required
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
      deprecated: false
      security:
        -
          Bearer: []
      operationId: createRecipe
      summary: Create a Recipe
      description: Creates a new instance of a `Recipe`.
    delete:
      tags:
        - /recipes
      responses:
        '204':
          description: Successful response - deletes all `Recipe`'s from array.
        '400':
          content:
            text/html:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad Request
        '401':
          content:
            text/html:
              examples:
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
      security:
        -
          Bearer: []
      operationId: deleteAllRecipes
      summary: Deletes All Recipes
      description: Deletes all instances of a `Recipe`.
  '/recipes/{id}':
    summary: Path used to manage the list of recipes.
    description: >-
      The REST endpoint/path used to edit and delete `Recipe` entities.  This path contains a `PUT` and
      `DELETE` operation to perform the edit and delete recipes, respectively.
    put:
      requestBody:
        description: A `Recipe` to be edited.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
            examples:
              Recipe:
                value:
                  content: some text
                  title: some text
        required: true
      tags:
        - '/recipes/:id'
      parameters:
        -
          name: id
          description: A recipe to be edited
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            text/html:
              examples:
                Recipe:
                  value:
                    id: 1
                    title: some text
                    content: some text
                    userId: 1
          description: OK
        '400':
          content:
            text/html:
              examples:
                Title and content are required:
                  value: Title and content are required
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad request (Title and content are required)
        '401':
          content:
            text/html:
              examples:
                Unauthorized:
                  value: Unauthorized
                Invalid session:
                  value: Invalid session
                Missing authorization header:
                  value: Missing authorization header
          description: Unauthorized
        '404':
          content:
            text/html:
              examples:
                Recipe not found:
                  value: Recipe not found
          description: Recipe ID not found.
      deprecated: false
      security:
        -
          Bearer: []
      operationId: editRecipe
      summary: Edit Recipe
      description: Updates an instance of `Recipe`.
    delete:
      tags:
        - '/recipes/:id'
      parameters:
        -
          name: id
          description: A recipe to be deleted
          schema:
            type: integer
          in: path
          required: true
      responses:
        '204':
          description: Successful response - deletes specific `Recipe` from array.
        '400':
          content:
            text/html:
              examples:
                Invalid authorization header:
                  value: Invalid authorization header
          description: Bad request
        '401':
          content:
            text/html:
              examples:
                Unauthorized:
                  value: Unauthorized
                Missing authorization header:
                  value: Missing authorization header
                Invalid session:
                  value: Invalid session
          description: Unauthorized
        '404':
          content:
            text/html:
              examples:
                Recipe not found:
                  value: Recipe not found
          description: Recipe ID not found in database.
      deprecated: false
      security:
        -
          Bearer: []
      operationId: deleteRecipe
      summary: Delete Recipe
      description: Deletes an instance of a `Recipe`.
    parameters:
      -
        name: id
        in: path
        required: true
components:
  schemas:
    User:
      title: Root Type for User
      description: ''
      type: object
      properties:
        id:
          format: int32
          description: A field that represents the unique identifier of the user.
          type: integer
          example: '12345'
        email:
          description: A field that represents the email address of the user.
          type: string
          example: john.doe@example.com
        password:
          description: A field that represents the password of the user.
          type: string
          example: MySecurePassword123
      example:
        id: 3
        email: test.test@gmail.com
        password: Passw0rd
    Session:
      title: Root Type for Session
      description: ''
      type: object
      properties:
        id:
          type: string
        userId:
          format: int32
          type: integer
      example:
        id: 330d67de-432b-48ce-a617-0bd6774ebcee
        userId: 1
    Credentials:
      title: Root Type for Credentials
      description: ''
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: admin
        password: admin
    Recipe:
      title: Root Type for Recipe
      description: ''
      type: object
      properties:
        content:
          type: string
        id:
          format: int32
          type: integer
        title:
          type: string
        userId:
          format: int32
          type: integer
      example:
        content: This is the content of example recipe 1
        id: 1
        title: Example 1
        userId: 1
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
tags:
  -
    name: /users
    description: ''
  -
    name: /sessions
    description: ''
  -
    name: /recipes
    description: ''
  -
    name: '/recipes/:id'
    description: ''